requirements.txt
# Core
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

# Embeddings
sentence-transformers>=2.2
torch>=2.0

# Vector search
faiss-cpu>=1.7

# Hybrid search
rank-bm25>=0.2

# RAG (optional — works without these)
# openai>=1.0

# Visualization
matplotlib>=3.7
seaborn>=0.12
wordcloud>=1.9

# UMAP (optional)
umap-learn>=0.5

# NLP utilities
textblob>=0.17

# Demo
streamlit>=1.28

# Testing
pytest>=7.4
flake8>=6.1

# Utilities
tqdm>=4.65
python-dotenv>=1.0
